+++ core/src/loader/mod.rs
@@ -1,1 +1,4 @@
 pub mod gnu_units;
+mod load;
+
+pub(crate) use load::load_defs;
